convert-pdf-to-image/requirements.txt
PyMuPDF==1.23.8
Pillow==10.1.0 
